requirements.txt
discord.py-self>=1.0.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
opuslib>=1.0.0